requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
httpie==2.0.0
idna==2.9
mysqlclient==1.4.6
Pygments==2.6.1
PyJWT==1.7.1
python-decouple==3.3
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
whitenoise==5.0.1
